+++ TRTC/src/core/canvas.h
@@ -1,6 +1,7 @@
 #ifndef TRTC_CANVAS_H
 #define TRTC_CANVAS_H
 
+#include <stdint.h>
 #include <stdbool.h>
 
 #include "color.h"
@@ -13,6 +14,6 @@ extern void canvas_set_pixel(int _x, int _y, struct Color _c);
 
 extern int canvas_get_pixel_count(void);
 
-extern void canvas_get_as_int_array(int _colors[]);
+extern void canvas_get_as_int_array(uint8_t _colors[]);
 
 #endif // TRTC_CANVAS_H
